requirements.txt
python-dotenv>=1.0.0
aiohttp>=3.8.0
playwright>=1.30.0
python-telegram-bot>=20.0
python-json-logger>=2.0.4
pymupdf>=1.23.0
